+++ CHANGELOG.md
@@ -45,9 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/cvat-ai/datumaro/pull/46>)
 - Skeleton support in datumaro format
   (<https://github.com/cvat-ai/datumaro/pull/47>)
-- Support for YOLOv8 formats
+- Support for Ultralytics YOLO formats
   (<https://github.com/cvat-ai/datumaro/pull/50>)
-- Support for YOLOv8 Classification format
+- Support for Ultralytics YOLO Classification format
   (<https://github.com/cvat-ai/datumaro/pull/59>)
 
 ### Changed